requirements.txt
amqp==5.0.2
ansible==2.10.6
ansible-base==2.10.5
appdirs==1.4.3
APScheduler==3.7.0
asgiref==3.2.10
bcrypt==3.2.0
billiard==3.6.3.0
casbin==0.16.0
celery==5.0.5
certifi==2020.4.5.1
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
dataclasses==0.8
distlib==0.3.0
Django==3.1.2
django-apscheduler==0.5.2
django-cors-headers==3.4.0
django-filter==2.4.0
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-mptt==0.11.0
django-mysql==3.10.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-datatables==0.5.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
Faker==6.0.0
fbtftp==0.2
fernet==1.0.1
filelock==3.0.12
gevent==21.1.2
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.8
paramiko==2.7.2
Pillow==8.1.0
pipenv==2018.11.26
prompt-toolkit==3.0.9
pyaes==1.6.1
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simpleeval==0.9.10
six==1.14.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.0.18
virtualenv-clone==0.5.4
wcwidth==0.2.5
xlrd==2.0.1
zipp==3.1.0
zmail==0.2.8
zope.event==4.5.0
zope.interface==5.2.0
